+++ judge-stack/src/main/java/com/dwarfeng/judge/stack/handler/SupportHandler.java
@@ -1,0 +1,27 @@
+package com.dwarfeng.judge.stack.handler;
+
+import com.dwarfeng.subgrade.stack.exception.HandlerException;
+import com.dwarfeng.subgrade.stack.handler.Handler;
+
+/**
+ * 支持处理器。
+ *
+ * @author DwArFeng
+ * @since 1.7.0
+ */
+public interface SupportHandler extends Handler {
+
+    /**
+     * 重置驱动器。
+     *
+     * @throws HandlerException 处理器异常。
+     */
+    void resetDriver() throws HandlerException;
+
+    /**
+     * 重置判断器。
+     *
+     * @throws HandlerException 处理器异常。
+     */
+    void resetJudger() throws HandlerException;
+}